refactor: 💡 update packages, small backend QOL improvements
Improvements include hiding seasons from shows that have no episodes,
hiding movies from collections that don't have a release date yet,
putting similar prisma calls in a transaction, small performance
upgrades

✅ Closes: BUC-70

+++ src/server/api/routers/movieList.ts
@@ -1,40 +1,11 @@
-import type { createTRPCContext } from '../trpc';
 import { createTRPCRouter, protectedProcedure } from '../trpc';
 import { z } from 'zod';
 import { zNewMovieSchema } from '~/schemas/listSchemas';
 import { TRPCError } from '@trpc/server';
-import type { Collection, List, User } from '@prisma/client';
+import type { Collection } from '@prisma/client';
 import { checkAndUpdateCollection, checkAndUpdateMovie } from '../../tmdbApi';
 import { getMovie, transformAPIMovie } from '~/server/TMDB/getMovie';
-
-export const checkAccess = (
-  ctx: Awaited<ReturnType<typeof createTRPCContext>>,
-  list: Partial<List & { collaborators: Pick<User, 'id'>[] }>,
-) =>
-  list.ownerId === ctx.auth.userId ||
-  list.collaborators?.some((c) => c.id === ctx.auth.userId);
-
-export const checkIfExistsAndAccess = (
-  ctx: Awaited<ReturnType<typeof createTRPCContext>>,
-  list:
-    | Partial<List & { collaborators: Pick<User, 'id'>[] }>
-    | undefined
-    | null,
-): list is Partial<List & { collaborators: Pick<User, 'id'>[] }> => {
-  if (!list)
-    throw new TRPCError({
-      code: 'NOT_FOUND',
-      message: 'That list cannot be found.',
-    });
-
-  if (!checkAccess(ctx, list))
-    throw new TRPCError({
-      code: 'UNAUTHORIZED',
-      message: 'You are not allowed to update this entity.',
-    });
-
-  return true;
-};
+import { checkIfExistsAndAccess } from '~/server/utils/checkIfExistsAndAccess';
 
 export const movieListRouter = createTRPCRouter({
   setMovieWatched: protectedProcedure
@@ -48,27 +19,35 @@ export const movieListRouter = createTRPCRouter({
     .mutation(async ({ ctx, input }) => {
       const list = await ctx.prisma.list.findUnique({
         where: { id: input.listId },
-        select: { ownerId: true, collaborators: { select: { id: true } } },
+        select: {
+          ownerId: true,
+          type: true,
+          collaborators: { select: { id: true } },
+        },
       });
 
-      checkIfExistsAndAccess(ctx, list);
+      checkIfExistsAndAccess(ctx, list, 'MOVIE');
 
-      await ctx.prisma.list.update({
-        where: { id: input.listId },
-        data: { updatedAt: new Date() },
-      });
-
-      if (input.checked) {
-        return ctx.prisma.checkedMovie.create({
-          data: {
-            list: { connect: { id: input.listId } },
-            movie: { connect: { id: input.id } },
-          },
+      return await ctx.prisma.$transaction(async (prisma) => {
+        await prisma.list.update({
+          where: { id: input.listId },
+          data: { updatedAt: new Date() },
         });
-      }
 
-      return ctx.prisma.checkedMovie.delete({
-        where: { movieId_listId: { listId: input.listId, movieId: input.id } },
+        if (input.checked) {
+          return await prisma.checkedMovie.create({
+            data: {
+              list: { connect: { id: input.listId } },
+              movie: { connect: { id: input.id } },
+            },
+          });
+        } else {
+          return await prisma.checkedMovie.delete({
+            where: {
+              movieId_listId: { listId: input.listId, movieId: input.id },
+            },
+          });
+        }
       });
     }),
   setCollectionChecked: protectedProcedure
@@ -84,42 +63,43 @@ export const movieListRouter = createTRPCRouter({
         where: { id: input.listId },
         select: {
           ownerId: true,
+          type: true,
           collaborators: { select: { id: true } },
           collections: { where: { id: input.id }, include: { movies: true } },
         },
       });
 
-      if (
-        !checkIfExistsAndAccess(ctx, list) ||
-        !list.collections[0]?.movies.length
-      )
-        return;
+      checkIfExistsAndAccess(ctx, list, 'MOVIE');
 
-      await ctx.prisma.list.update({
-        where: { id: input.listId },
-        data: { updatedAt: new Date() },
-      });
-
-      if (input.checked) {
-        const movies = list.collections[0].movies.map((movie) => ({
-          id: movie.id,
-        }));
-
-        return ctx.prisma.checkedMovie.createMany({
-          data: movies.map((movie) => ({
-            listId: input.listId,
-            movieId: movie.id,
-          })),
+      return await ctx.prisma.$transaction(async (prisma) => {
+        await prisma.list.update({
+          where: { id: input.listId },
+          data: { updatedAt: new Date() },
         });
-      }
 
-      return ctx.prisma.checkedMovie.deleteMany({
-        where: {
-          listId: input.listId,
-          movieId: {
-            in: list.collections[0].movies.map((movie) => movie.id),
-          },
-        },
+        if (!list.collections[0]?.movies.length)
+          throw new TRPCError({
+            code: 'NOT_FOUND',
+            message: 'That collection does not exist.',
+          });
+
+        if (input.checked) {
+          return ctx.prisma.checkedMovie.createMany({
+            data: list.collections[0].movies.map((movie) => ({
+              listId: input.listId,
+              movieId: movie.id,
+            })),
+          });
+        } else {
+          return ctx.prisma.checkedMovie.deleteMany({
+            where: {
+              listId: input.listId,
+              movieId: {
+                in: list.collections[0].movies.map((movie) => movie.id),
+              },
+            },
+          });
+        }
       });
     }),
   createMovie: protectedProcedure
@@ -128,22 +108,16 @@ export const movieListRouter = createTRPCRouter({
       const list = await ctx.prisma.list.findUnique({
         where: { id: input.listId },
         select: {
+          _count: { select: { movies: { where: { id: input.externalId } } } },
           ownerId: true,
           collaborators: { select: { id: true } },
           type: true,
-          movies: { select: { id: true }, where: { id: input.externalId } },
         },
       });
 
-      if (!checkIfExistsAndAccess(ctx, list)) return;
+      checkIfExistsAndAccess(ctx, list, 'MOVIE');
 
-      if (list.type !== 'MOVIE')
-        throw new TRPCError({
-          code: 'BAD_REQUEST',
-          message: 'You can only add movie list items to this list.',
-        });
-
-      if (list.movies.length) {
+      if (list._count.movies) {
         throw new TRPCError({
           code: 'CONFLICT',
           message: 'That movie is already in this list.',
@@ -156,11 +130,6 @@ export const movieListRouter = createTRPCRouter({
 
       if (!movie) {
         const tmdbMovie = await getMovie(input.externalId);
-        if (!tmdbMovie)
-          throw new TRPCError({
-            code: 'NOT_FOUND',
-            message: 'Something went wrong finding that movie on TMDB.',
-          });
 
         movie = await ctx.prisma.movie.create({
           data: {
@@ -185,25 +154,18 @@ export const movieListRouter = createTRPCRouter({
       const list = await ctx.prisma.list.findUnique({
         where: { id: input.listId },
         select: {
+          _count: {
+            select: { collections: { where: { id: input.externalId } } },
+          },
           ownerId: true,
           collaborators: { select: { id: true } },
           type: true,
-          collections: {
-            where: { id: input.externalId },
-            select: { id: true },
-          },
         },
       });
 
-      if (!checkIfExistsAndAccess(ctx, list)) return;
+      checkIfExistsAndAccess(ctx, list, 'MOVIE');
 
-      if (list.type !== 'MOVIE')
-        throw new TRPCError({
-          code: 'BAD_REQUEST',
-          message: 'You can only add movie list items to this list.',
-        });
-
-      if (list.collections.length) {
+      if (list._count.collections) {
         throw new TRPCError({
           code: 'CONFLICT',
           message: 'That collection is already in this list.',
@@ -217,14 +179,11 @@ export const movieListRouter = createTRPCRouter({
         });
 
       if (!collection) {
-        collection = await checkAndUpdateCollection(
-          ctx,
-          collection ?? {
-            id: input.externalId,
-            updatedAt: new Date('01-01-2000'),
-            etag: '',
-          },
-        );
+        collection = await checkAndUpdateCollection(ctx, {
+          id: input.externalId,
+          updatedAt: new Date('01-01-2000'),
+          etag: '',
+        });
       }
 
       if (!collection)
@@ -248,24 +207,31 @@ export const movieListRouter = createTRPCRouter({
     .mutation(async ({ ctx, input }) => {
       const list = await ctx.prisma.list.findUnique({
         where: { id: input.listId },
-        select: { ownerId: true, collaborators: { select: { id: true } } },
-      });
-
-      if (!checkIfExistsAndAccess(ctx, list)) return;
-
-      await ctx.prisma.checkedMovie.deleteMany({
-        where: {
-          AND: [{ movieId: input.id }, { listId: input.listId }],
+        select: {
+          ownerId: true,
+          type: true,
+          collaborators: { select: { id: true } },
         },
       });
 
-      return ctx.prisma.list.update({
-        where: { id: input.listId },
-        data: {
-          movies: { disconnect: { id: input.id } },
-          updatedAt: new Date(),
-        },
-      });
+      checkIfExistsAndAccess(ctx, list, 'MOVIE');
+
+      const [updatedList] = await ctx.prisma.$transaction([
+        ctx.prisma.list.update({
+          where: { id: input.listId },
+          data: {
+            movies: { disconnect: { id: input.id } },
+            updatedAt: new Date(),
+          },
+        }),
+        ctx.prisma.checkedMovie.deleteMany({
+          where: {
+            AND: [{ movieId: input.id }, { listId: input.listId }],
+          },
+        }),
+      ]);
+
+      return updatedList;
     }),
   deleteCollection: protectedProcedure
     .input(z.object({ id: z.number(), listId: z.string() }))
@@ -274,6 +240,7 @@ export const movieListRouter = createTRPCRouter({
         where: { id: input.listId },
         select: {
           ownerId: true,
+          type: true,
           collaborators: { select: { id: true } },
           collections: {
             select: { movies: { select: { id: true } } },
@@ -282,7 +249,7 @@ export const movieListRouter = createTRPCRouter({
         },
       });
 
-      if (!checkIfExistsAndAccess(ctx, list)) return;
+      checkIfExistsAndAccess(ctx, list, 'MOVIE');
 
       if (!list.collections[0])
         throw new TRPCError({
@@ -291,21 +258,24 @@ export const movieListRouter = createTRPCRouter({
             "The collection you're requesting to delete cannot be found.",
         });
 
-      await ctx.prisma.checkedMovie.deleteMany({
-        where: {
-          AND: [
-            { movieId: { in: list.collections[0].movies.map((m) => m.id) } },
-            { listId: input.listId },
-          ],
-        },
-      });
+      const [updatedList] = await ctx.prisma.$transaction([
+        ctx.prisma.list.update({
+          where: { id: input.listId },
+          data: {
+            collections: { disconnect: { id: input.id } },
+            updatedAt: new Date(),
+          },
+        }),
+        ctx.prisma.checkedMovie.deleteMany({
+          where: {
+            AND: [
+              { movieId: { in: list.collections[0].movies.map((m) => m.id) } },
+              { listId: input.listId },
+            ],
+          },
+        }),
+      ]);
 
-      return ctx.prisma.list.update({
-        where: { id: input.listId },
-        data: {
-          collections: { disconnect: { id: input.id } },
-          updatedAt: new Date(),
-        },
-      });
+      return updatedList;
     }),
 });
